requirements.txt
httpx>=0.27.0
lxml>=5.1.0
pydantic>=2.9.0
python-dateutil>=2.9.0

